01_Ai_Agents_First/02_Open Router/pyproject.toml
[project]
name = "02-open-router"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.0.19",
    "python-dotenv>=1.1.0",
]
